+++ SingleThrift/tailwind.config.js
@@ -3,6 +3,7 @@
 module.exports = {
   content: [
     './App.{js,jsx,ts,tsx}',
+    './Appholder.{js,jsx,ts,tsx}',
     './components/**/*.{js,jsx,ts,tsx}',
     './screens/**/*.{js,jsx,ts,tsx}',
     './navigators/**/*.{js,jsx,ts,tsx}',
